pyproject.toml
[tool.poetry]
name = "gpt-auto-applier"
version = "0.1.0"
description = ""
authors = ["John Walz <johnwalz97@gmail.com>"]
readme = "README.md"
packages = [{include = "gpt_auto_applier"}]

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.12.0"
requests = "^2.28.2"
openai = "^0.27.4"
selenium = "^4.8.3"
python-dotenv = "^1.0.0"
pypdf2 = "^3.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
ruff = "^0.0.260"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
